Implementación REST casi terminada

+++ modelo/src/main/java/com/ipartek/formacion/modelo/pojo/Coche.java
@@ -1,12 +1,12 @@
 package com.ipartek.formacion.modelo.pojo;
 
 public class Coche {
-	private Long id;
+	private int id;
 	private String matricula;
 	private String modelo;
 	private Integer km;
 	
-	public Coche(Long id, String matricula, String modelo, Integer km) {
+	public Coche(int id, String matricula, String modelo, Integer km) {
 		this();
 		setId(id);
 		setMatricula(matricula);
@@ -16,17 +16,17 @@ public class Coche {
 
 	public Coche() {
 		super();
-		this.id=-1l;
+		this.id=-1;
 		this.matricula="";
 		this.modelo="";
 		this.km=0;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
